CompareTo non-generic method
